// src/types/regularpurchase/regularPurchase.tsx
export type RegularPurchase = {
    created_at?: string;
    img_url: string;
    last_order_created_at?: string;
    order_created_at?: string;
    product_code: string;
    product_name: string;
    product_subscribe_id: number;
    qty: number;
    status: number;
    subscribe_kikan: number;
    subscribe_kikan_id: number;
    subscribe_value: number;
    title: string;
    user_code: string;
    user_email: string;
    user_name: string;
    user_name_kana: string;
    user_phone: string; 
}